Depth based positioning, Abstract methods, Clears up basic UI methods/fields, Own file for BasicCheckBox
